Move audio playing into audioSource tag
