+++ Assets/Scripts/UnityInterface/AudioSourceTag.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3014bfea922e6304db0cc64a8da229df
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
